backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Background Tasks
celery[redis]==5.4.0
redis==5.2.0

# LLM & RAG
openai>=1.50.0
llama-index-core==0.12.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-parse>=0.5.0

# Vector Database
qdrant-client>=1.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
aiofiles>=24.0.0
google-generativeai>=0.8.0